+++ core-ng-test/src/main/java/core/framework/test/redis/MockRedisList.java
@@ -35,9 +35,15 @@ public final class MockRedisList implements RedisList {
     }
 
     @Override
-    public List<String> range(String key, int start, int end) {
+    public List<String> range(String key, long start, long end) {
         var value = store.get(key);
         if (value == null) return List.of();
-        return List.copyOf(value.list());
+        List<String> list = value.list();
+        int size = list.size();
+        int startIndex = start < 0 ? 0 : (int) start;
+        if (startIndex > size) startIndex = size;
+        int endIndex = end < 0 ? (int) end + size : (int) end;
+        if (endIndex >= size) endIndex = size - 1;
+        return List.copyOf(list.subList(startIndex, endIndex + 1));
     }
 }