+++ core-ng/src/main/java/core/framework/impl/redis/RedisEncodings.java
@@ -14,7 +14,7 @@ class RedisEncodings {
     static {
         for (int i = 0; i < 256; i++) {
             String text = Integer.toString(i);
-            INT_BYTES_CACHE[i] = bytes(text);
+            INT_BYTES_CACHE[i] = Strings.bytes(text);
         }
     }
 
@@ -24,19 +24,11 @@ class RedisEncodings {
     }
 
     static byte[] encode(long value) {
-        if (value >= 0 && value <= 255) {
+        if (value >= 0 && value < 256) {
             return INT_BYTES_CACHE[(int) value];
         }
         String text = Long.toString(value);
-        return bytes(text);
-    }
-
-    static byte[] encode(int value) {
-        if (value >= 0 && value <= 255) {
-            return INT_BYTES_CACHE[value];
-        }
-        String text = Integer.toString(value);
-        return bytes(text);
+        return Strings.bytes(text);
     }
 
     static byte[][] encode(String key, String... values) {
@@ -46,6 +38,16 @@ class RedisEncodings {
         result[0] = encode(key);
         for (int i = 0; i < length; i++) {
             result[i + 1] = encode(values[i]);
+        }
+        return result;
+    }
+
+    static byte[][] encode(String... values) {
+        if (values.length == 0) throw new RedisException("values must not be empty");
+        int length = values.length;
+        byte[][] result = new byte[length][];
+        for (int i = 0; i < length; i++) {
+            result[i] = encode(values[i]);
         }
         return result;
     }
@@ -76,14 +78,5 @@ class RedisEncodings {
     static String decode(byte[] value) {
         if (value == null) return null;
         return new String(value, Charsets.UTF_8);
-    }
-
-    private static byte[] bytes(String text) {
-        char[] chars = text.toCharArray();
-        byte[] bytes = new byte[chars.length];
-        for (int i = 0; i < chars.length; i++) {
-            bytes[i] = (byte) chars[i];
-        }
-        return bytes;
     }
 }